// user/src/pattern.enum.ts
export enum Pattern {
  REGISTER = "USER_REGISTER",
  LOGIN = "USER_LOGIN",
  UPDATE = "USER_UPDATE",
  GET_USER = "USER_GET_USER",
  CHANGE_PASSWORD_REQUEST = "USER_CHANGE_PASSWORD_REQUEST",
  CHANGE_PASSWORD = "USER_CHANGE_PASSWORD",
  GRANT_ADMIN = "USER_GRANT_ADMIN",
  TAKE_ADMIN = "USER_TAKE_ADMIN",
  BAN_USER = "USER_BAN_USER",
  VALIDATE_TOKEN = "USER_VALIDATE_TOKEN",
  GET_USER_BY_ID = "USER_GET_USER_BY_ID",
  ADD_USER_TEAM = "USER_ADD_TEAM",
  REMOVE_USER_TEAM = "USER_REMOVE_TEAM",
  ADD_USER_TODO = "USER_ADD_TODO",
  REMOVE_USER_TODO = "USER_REMOVE_TODO",
  GET_TEAM_BY_ID = "TEAM_GET_TEAM_BY_ID",
  GET_TODO_BY_ID = "TODO_GET_TODO_BY_ID",
}